OLH-2620: Move AccountInterventionsService to shared
Along with the tests and supporting entities.

We'll need to use this service in the account management API as to
check for interventions when a user is changing their details.
